+++ bsdploy/tests/test_bootstrap_mfsbsd.py
@@ -14,11 +14,23 @@ def bootstrap(env_mock, environ_mock, monkeypatch, put_mock, run_mock, tempdir, 
     return bootstrap
 
 
+def create_ssh_host_keys(tempdir):
+    tempdir['bootstrap-files/ssh_host_dsa_key'].fill('dsa')
+    tempdir['bootstrap-files/ssh_host_dsa_key.pub'].fill('dsa.pub')
+    tempdir['bootstrap-files/ssh_host_ecdsa_key'].fill('ecdsa')
+    tempdir['bootstrap-files/ssh_host_ecdsa_key.pub'].fill('ecdsa.pub')
+    tempdir['bootstrap-files/ssh_host_key'].fill('rsa1')
+    tempdir['bootstrap-files/ssh_host_key.pub'].fill('rsa1.pub')
+    tempdir['bootstrap-files/ssh_host_rsa_key'].fill('rsa')
+    tempdir['bootstrap-files/ssh_host_rsa_key.pub'].fill('rsa.pub')
+
+
 def test_bootstrap_ask_to_continue(bootstrap, capsys, run_mock, tempdir, yesno_mock):
     format_info = dict(
         bsdploy_path=bsdploy_path,
         tempdir=tempdir.directory)
-    tempdir['etc/authorized_keys'].fill('id_dsa')
+    tempdir['bootstrap-files/authorized_keys'].fill('id_dsa')
+    create_ssh_host_keys(tempdir)
     run_mock.expected = [
         ('mount', {}, default_mounts),
         ('test -e /dev/cd0 && mount_cd9660 /dev/cd0 /cdrom || true', {}, '\n'),
@@ -36,10 +48,18 @@ def test_bootstrap_ask_to_continue(bootstrap, capsys, run_mock, tempdir, yesno_m
         "",
         "Using these local files for bootstrapping:",
         "%(bsdploy_path)s/bootstrap-files/FreeBSD.conf -(template:False)-> /mnt/usr/local/etc/pkg/repos/FreeBSD.conf" % format_info,
-        "%(tempdir)s/etc/authorized_keys -(template:False)-> /mnt/root/.ssh/authorized_keys" % format_info,
+        "%(tempdir)s/bootstrap-files/authorized_keys -(template:False)-> /mnt/root/.ssh/authorized_keys" % format_info,
         "%(bsdploy_path)s/bootstrap-files/make.conf -(template:False)-> /mnt/etc/make.conf" % format_info,
         "%(bsdploy_path)s/bootstrap-files/pkg.conf -(template:False)-> /mnt/usr/local/etc/pkg.conf" % format_info,
         "%(bsdploy_path)s/bootstrap-files/rc.conf -(template:True)-> /mnt/etc/rc.conf" % format_info,
+        "%(tempdir)s/bootstrap-files/ssh_host_dsa_key -(template:False)-> /mnt/etc/ssh/ssh_host_dsa_key" % format_info,
+        "%(tempdir)s/bootstrap-files/ssh_host_dsa_key.pub -(template:False)-> /mnt/etc/ssh/ssh_host_dsa_key.pub" % format_info,
+        "%(tempdir)s/bootstrap-files/ssh_host_ecdsa_key -(template:False)-> /mnt/etc/ssh/ssh_host_ecdsa_key" % format_info,
+        "%(tempdir)s/bootstrap-files/ssh_host_ecdsa_key.pub -(template:False)-> /mnt/etc/ssh/ssh_host_ecdsa_key.pub" % format_info,
+        "%(tempdir)s/bootstrap-files/ssh_host_key -(template:False)-> /mnt/etc/ssh/ssh_host_key" % format_info,
+        "%(tempdir)s/bootstrap-files/ssh_host_key.pub -(template:False)-> /mnt/etc/ssh/ssh_host_key.pub" % format_info,
+        "%(tempdir)s/bootstrap-files/ssh_host_rsa_key -(template:False)-> /mnt/etc/ssh/ssh_host_rsa_key" % format_info,
+        "%(tempdir)s/bootstrap-files/ssh_host_rsa_key.pub -(template:False)-> /mnt/etc/ssh/ssh_host_rsa_key.pub" % format_info,
         "%(bsdploy_path)s/bootstrap-files/sshd_config -(template:False)-> /mnt/etc/ssh/sshd_config" % format_info,
         "",
         "The following files will be downloaded on the host during bootstrap:",
@@ -58,8 +78,9 @@ def test_bootstrap_ask_to_continue(bootstrap, capsys, run_mock, tempdir, yesno_m
         "Continue?"]
 
 
-def test_bootstrap_no_newline_at_end_of_rc_conf(bootstrap, capsys, run_mock, tempdir):
-    tempdir['etc/authorized_keys'].fill('id_dsa')
+def test_bootstrap_no_newline_at_end_of_rc_conf(bootstrap, capsys, local_mock, run_mock, tempdir):
+    tempdir['bootstrap-files/authorized_keys'].fill('id_dsa')
+    create_ssh_host_keys(tempdir)
     tempdir['bootstrap-files/rc.conf'].fill('foo')
     run_mock.expected = [
         ('mount', {}, default_mounts),
@@ -83,14 +104,23 @@ def test_bootstrap(bootstrap, put_mock, run_mock, tempdir, yesno_mock):
     format_info = dict(
         bsdploy_path=bsdploy_path,
         tempdir=tempdir.directory)
-    tempdir['etc/authorized_keys'].fill('id_dsa')
+    tempdir['bootstrap-files/authorized_keys'].fill('id_dsa')
+    create_ssh_host_keys(tempdir)
     put_mock.expected = [
         (("%(bsdploy_path)s/bootstrap-files/FreeBSD.conf" % format_info, '/mnt/usr/local/etc/pkg/repos/FreeBSD.conf'), {'mode': None}),
-        (("%(tempdir)s/etc/authorized_keys" % format_info, '/mnt/root/.ssh/authorized_keys'), {'mode': None}),
+        (("%(tempdir)s/bootstrap-files/authorized_keys" % format_info, '/mnt/root/.ssh/authorized_keys'), {'mode': None}),
         (("%(bsdploy_path)s/bootstrap-files/make.conf" % format_info, '/mnt/etc/make.conf'), {'mode': None}),
         (("%(bsdploy_path)s/bootstrap-files/pkg.conf" % format_info, '/mnt/usr/local/etc/pkg.conf'), {'mode': None}),
         # put from upload_template
         ((object, '/mnt/etc/rc.conf'), {'mode': None}),
+        (("%(tempdir)s/bootstrap-files/ssh_host_dsa_key" % format_info, '/mnt/etc/ssh/ssh_host_dsa_key'), {'mode': 0600}),
+        (("%(tempdir)s/bootstrap-files/ssh_host_dsa_key.pub" % format_info, '/mnt/etc/ssh/ssh_host_dsa_key.pub'), {'mode': 0644}),
+        (("%(tempdir)s/bootstrap-files/ssh_host_ecdsa_key" % format_info, '/mnt/etc/ssh/ssh_host_ecdsa_key'), {'mode': 0600}),
+        (("%(tempdir)s/bootstrap-files/ssh_host_ecdsa_key.pub" % format_info, '/mnt/etc/ssh/ssh_host_ecdsa_key.pub'), {'mode': 0644}),
+        (("%(tempdir)s/bootstrap-files/ssh_host_key" % format_info, '/mnt/etc/ssh/ssh_host_key'), {'mode': 0600}),
+        (("%(tempdir)s/bootstrap-files/ssh_host_key.pub" % format_info, '/mnt/etc/ssh/ssh_host_key.pub'), {'mode': 0644}),
+        (("%(tempdir)s/bootstrap-files/ssh_host_rsa_key" % format_info, '/mnt/etc/ssh/ssh_host_rsa_key'), {'mode': 0600}),
+        (("%(tempdir)s/bootstrap-files/ssh_host_rsa_key.pub" % format_info, '/mnt/etc/ssh/ssh_host_rsa_key.pub'), {'mode': 0644}),
         (("%(bsdploy_path)s/bootstrap-files/sshd_config" % format_info, '/mnt/etc/ssh/sshd_config'), {'mode': None}),
     ]
     run_mock.expected = [
@@ -115,11 +145,6 @@ def test_bootstrap(bootstrap, put_mock, run_mock, tempdir, yesno_mock):
         ('chroot /mnt pkg2ng', {}, ''),
         ('chroot /mnt pkg install python27', {}, ''),
         ('echo autoboot_delay=-1 >> /mnt/boot/loader.conf', {}, ''),
-        ("ssh-keygen -t dsa -f /mnt/etc/ssh/ssh_host_dsa_key -N ''", {}, ''),
-        ("ssh-keygen -t ecdsa -f /mnt/etc/ssh/ssh_host_ecdsa_key -N ''", {}, ''),
-        ("ssh-keygen -t rsa1 -b 1024 -f /mnt/etc/ssh/ssh_host_key -N ''", {}, ''),
-        ("ssh-keygen -t rsa -f /mnt/etc/ssh/ssh_host_rsa_key -N ''", {}, ''),
-        ('ssh-keygen -lf /mnt/etc/ssh/ssh_host_rsa_key', {}, ''),
         ('reboot', {}, '')]
     yesno_mock.expected = [
         ("\nContinuing will destroy the existing data on the following devices:\n    ada0\n\nContinue?", True)]