+++ CHANGES.rst
@@ -1,6 +1,9 @@
 1.0b5 - unreleased
 ==================
 
+- automatically set env.shell for fabric scripts.
+- generate ssh host keys locally during bootstrap if possible.
+- set ``fingerprint`` option for ezjail master automatically if a ssh host key exists locally.
 
 
 1.0b4 - 2014-07-08